requirements.txt
streamlit
rembg
Pillow
numpy
opencv-python-headless
